@@ -9,21 +9,22 @@
 
 #include "../State.hpp"
 #include "../util_game.hpp"
-
+typedef std::shared_ptr<unordered_map<string ,string>> dictionary;
 class Policy{
+
 public:
     int max_speed;
     string name;
     bool out_budget;
     bool is_wall;
     int D=2;
-    unordered_map<string,string>* infoDict;
+    dictionary infoDict;
     bool evalPolicy;
     string home;
     string id_agent;
     unordered_map<int,Point*>* hashActionMap;
     vector<Policy*> tran;
-    Policy(string name_policy,int max_speed_agent,string agentID,string &_home,unordered_map<string,string> *info)
+    Policy(string name_policy,int max_speed_agent,string agentID,string &_home,dictionary info)
     :max_speed(max_speed_agent),evalPolicy(false),infoDict(info),home(_home){
         this->name=std::move(name_policy);
         this->is_wall=false;
